+++ examples/tensormap_and_ringbuffer/paged_attention/kernels/orchestration/paged_attention_orch.cpp
@@ -71,9 +71,9 @@ void aicpu_orchestration_entry(PTO2Runtime* rt, uint64_t* args, int arg_count) {
     int64_t* host_config = (int64_t*)(uintptr_t)args[6];
 
     // Extract sizes (next 7 args after pointers)
-    size_t query_size = (size_t)args[7];
-    size_t key_cache_size = (size_t)args[8];
-    size_t value_cache_size = (size_t)args[9];
+    uint64_t query_size = (uint64_t)args[7];
+    uint64_t key_cache_size = (uint64_t)args[8];
+    uint64_t value_cache_size = (uint64_t)args[9];
 
     // Extract config parameters
     uint64_t batch = (uint64_t)(int)host_config[0];
