+++ src/platform/a2a3/host/device_runner.h
@@ -159,7 +159,7 @@ public:
      * @param bytes  Size of tensor in bytes
      * @return Device pointer on success, nullptr on failure
      */
-    void* allocate_tensor(size_t bytes);
+    void* allocate_tensor(uint64_t bytes);
 
     /**
      * Free device tensor memory
@@ -176,7 +176,7 @@ public:
      * @param bytes    Number of bytes to copy
      * @return 0 on success, error code on failure
      */
-    int copy_to_device(void* dev_ptr, const void* host_ptr, size_t bytes);
+    int copy_to_device(void* dev_ptr, const void* host_ptr, uint64_t bytes);
 
     /**
      * Copy data from device to host
@@ -186,7 +186,7 @@ public:
      * @param bytes    Number of bytes to copy
      * @return 0 on success, error code on failure
      */
-    int copy_from_device(void* host_ptr, const void* dev_ptr, size_t bytes);
+    int copy_from_device(void* host_ptr, const void* dev_ptr, uint64_t bytes);
 
     /**
      * Execute a runtime
@@ -303,7 +303,7 @@ public:
      * @param bin_size  Size of binary data in bytes
      * @return Device GM address of kernel on success, 0 on error
      */
-    uint64_t upload_kernel_binary(int func_id, const uint8_t* bin_data, size_t bin_size);
+    uint64_t upload_kernel_binary(int func_id, const uint8_t* bin_data, uint64_t bin_size);
 
     /**
      * Ensure device is set and streams are created (minimal initialization)
