+++ src/platform/a2a3/host/pto_runtime_c_api.cpp
@@ -20,7 +20,7 @@ extern "C" {
  */
 int init_runtime_impl(Runtime* runtime,
                     const uint8_t* orch_so_binary,
-                    size_t orch_so_size,
+                    uint64_t orch_so_size,
                     const char* orch_func_name,
                     uint64_t* func_args,
                     int func_args_count,
@@ -28,16 +28,16 @@ int init_runtime_impl(Runtime* runtime,
                     uint64_t* arg_sizes,
                     const int* kernel_func_ids,
                     const uint8_t* const* kernel_binaries,
-                    const size_t* kernel_sizes,
+                    const uint64_t* kernel_sizes,
                     int kernel_count);
 int validate_runtime_impl(Runtime* runtime);
 
 /* Forward declarations for device memory functions used in init_runtime */
-void* device_malloc(size_t size);
+void* device_malloc(uint64_t size);
 void device_free(void* dev_ptr);
-int copy_to_device(void* dev_ptr, const void* host_ptr, size_t size);
-int copy_from_device(void* host_ptr, const void* dev_ptr, size_t size);
-uint64_t upload_kernel_binary_wrapper(int func_id, const uint8_t* bin_data, size_t bin_size);
+int copy_to_device(void* dev_ptr, const void* host_ptr, uint64_t size);
+int copy_from_device(void* host_ptr, const void* dev_ptr, uint64_t size);
+uint64_t upload_kernel_binary_wrapper(int func_id, const uint8_t* bin_data, uint64_t bin_size);
 
 /* ===========================================================================
  */
@@ -45,11 +45,11 @@ uint64_t upload_kernel_binary_wrapper(int func_id, const uint8_t* bin_data, size
 /* ===========================================================================
  */
 
-size_t get_runtime_size(void) { return sizeof(Runtime); }
+uint64_t get_runtime_size(void) { return sizeof(Runtime); }
 
 int init_runtime(RuntimeHandle runtime,
                 const uint8_t* orch_so_binary,
-                size_t orch_so_size,
+                uint64_t orch_so_size,
                 const char* orch_func_name,
                 uint64_t* func_args,
                 int func_args_count,
@@ -57,7 +57,7 @@ int init_runtime(RuntimeHandle runtime,
                 uint64_t* arg_sizes,
                 const int* kernel_func_ids,
                 const uint8_t* const* kernel_binaries,
-                const size_t* kernel_sizes,
+                const uint64_t* kernel_sizes,
                 int kernel_count) {
     if (runtime == NULL) {
         return -1;
@@ -103,7 +103,7 @@ int init_runtime(RuntimeHandle runtime,
 /* ===========================================================================
  */
 
-void* device_malloc(size_t size) {
+void* device_malloc(uint64_t size) {
     try {
         DeviceRunner& runner = DeviceRunner::get();
         return runner.allocate_tensor(size);
@@ -124,7 +124,7 @@ void device_free(void* dev_ptr) {
     }
 }
 
-int copy_to_device(void* dev_ptr, const void* host_ptr, size_t size) {
+int copy_to_device(void* dev_ptr, const void* host_ptr, uint64_t size) {
     if (dev_ptr == NULL || host_ptr == NULL) {
         return -1;
     }
@@ -136,7 +136,7 @@ int copy_to_device(void* dev_ptr, const void* host_ptr, size_t size) {
     }
 }
 
-int copy_from_device(void* host_ptr, const void* dev_ptr, size_t size) {
+int copy_from_device(void* host_ptr, const void* dev_ptr, uint64_t size) {
     if (host_ptr == NULL || dev_ptr == NULL) {
         return -1;
     }
@@ -148,7 +148,7 @@ int copy_from_device(void* host_ptr, const void* dev_ptr, size_t size) {
     }
 }
 
-uint64_t upload_kernel_binary_wrapper(int func_id, const uint8_t* bin_data, size_t bin_size) {
+uint64_t upload_kernel_binary_wrapper(int func_id, const uint8_t* bin_data, uint64_t bin_size) {
     try {
         DeviceRunner& runner = DeviceRunner::get();
         return runner.upload_kernel_binary(func_id, bin_data, bin_size);
@@ -162,9 +162,9 @@ int launch_runtime(RuntimeHandle runtime,
     int block_dim,
     int device_id,
     const uint8_t* aicpu_binary,
-    size_t aicpu_size,
+    uint64_t aicpu_size,
     const uint8_t* aicore_binary,
-    size_t aicore_size) {
+    uint64_t aicore_size) {
     if (runtime == NULL) {
         return -1;
     }
@@ -218,7 +218,7 @@ int set_device(int device_id) {
 void record_tensor_pair(RuntimeHandle runtime,
                        void* host_ptr,
                        void* dev_ptr,
-                       size_t size) {
+                       uint64_t size) {
     if (runtime == NULL) {
         return;
     }
